decompose files. build failed
