+++ perception/autoware_map_based_prediction/include/autoware/map_based_prediction/path_generator/path_generator.hpp
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef MAP_BASED_PREDICTION__PATH_GENERATOR_HPP_
-#define MAP_BASED_PREDICTION__PATH_GENERATOR_HPP_
+#ifndef AUTOWARE_MAP_BASED_PREDICTION__PATH_GENERATOR__PATH_GENERATOR_HPP_
+#define AUTOWARE_MAP_BASED_PREDICTION__PATH_GENERATOR__PATH_GENERATOR_HPP_
 
-#include "map_based_prediction/data_structure.hpp"
+#include "autoware/map_based_prediction/data_structure.hpp"
 
 #include <Eigen/Eigen>
 #include <autoware_utils/system/time_keeper.hpp>
@@ -170,4 +170,4 @@ private:
 };
 }  // namespace autoware::map_based_prediction
 
-#endif  // MAP_BASED_PREDICTION__PATH_GENERATOR_HPP_
+#endif  // AUTOWARE_MAP_BASED_PREDICTION__PATH_GENERATOR__PATH_GENERATOR_HPP_
