+++ perception/autoware_map_based_prediction/include/autoware/map_based_prediction/predictor_vru/history.hpp
@@ -1,0 +1,20 @@
+// Copyright 2024 TIER IV, inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef AUTOWARE_MAP_BASED_PREDICTION__PREDICTOR_VRU__HISTORY_HPP_
+#define AUTOWARE_MAP_BASED_PREDICTION__PREDICTOR_VRU__HISTORY_HPP_
+
+#include "autoware/map_based_prediction/predictor_vru/predictor_vru.hpp"
+
+#endif  // AUTOWARE_MAP_BASED_PREDICTION__PREDICTOR_VRU__HISTORY_HPP_
