+++ perception/autoware_map_based_prediction/include/autoware/map_based_prediction/predictor_vru/predictor_vru.hpp
@@ -12,11 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef MAP_BASED_PREDICTION__PREDICTOR_VRU_HPP_
-#define MAP_BASED_PREDICTION__PREDICTOR_VRU_HPP_
+#ifndef AUTOWARE_MAP_BASED_PREDICTION__PREDICTOR_VRU__PREDICTOR_VRU_HPP_
+#define AUTOWARE_MAP_BASED_PREDICTION__PREDICTOR_VRU__PREDICTOR_VRU_HPP_
 
-#include "map_based_prediction/data_structure.hpp"
-#include "map_based_prediction/path_generator.hpp"
+#include "autoware/map_based_prediction/data_structure.hpp"
+#include "autoware/map_based_prediction/path_generator/path_generator.hpp"
 
 #include <autoware_utils/system/time_keeper.hpp>
 #include <rclcpp/rclcpp.hpp>
@@ -152,4 +152,4 @@ private:
 
 }  // namespace autoware::map_based_prediction
 
-#endif  // MAP_BASED_PREDICTION__PREDICTOR_VRU_HPP_
+#endif  // AUTOWARE_MAP_BASED_PREDICTION__PREDICTOR_VRU__PREDICTOR_VRU_HPP_
