+++ perception/autoware_map_based_prediction/test/map_based_prediction/test_path_generator.cpp
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "map_based_prediction/data_structure.hpp"
-#include "map_based_prediction/path_generator.hpp"
+#include "autoware/map_based_prediction/data_structure.hpp"
+#include "autoware/map_based_prediction/path_generator/path_generator.hpp"
 
 #include <gtest/gtest.h>
